make-update-zip.py: add zip signing
When the -c (--cert) and -k (--key) options are given, then the
signing action must succeed, otherwise the program fails. When not
given, a warning is displayed.

signapk.jar is generated when building AOSP and can be found at
$OUT_DIR/host/linux-x86/framework/signapk.jar. Alternatively, it can
be retrieved from
https://android.googlesource.com/platform/prebuilts/sdk (tools/lib/).
When not using git, but the web interface, note that due to a limitation
of Google's web interface you can only download a tgz or base64-encoded
file (https://code.google.com/p/gitiles/issues/detail?id=7).