make-update-zip.py: root make_files relative to system root
Avoids duplicating the common prefix of the full path and arcname.
This is possible because I now assume that the APKs are located in
a local path such as system/app/FooApp/FooApp.apk. That is necessary
because future support for symlinks require that the root is known
to be /system.